Update deprecated functions, bugfixes, macOS support
This patch updates the main code to reduce code duplication, and
replaced deprecated functions with newer ones.

Also adds a discussion button with comment count, and fixed opening
discussion-only threads.

A new "active" tab is also added, which was made possible with the code
refractor.

Added macOS build with proper entitlements for network. And tested on
iOS.